requirements.txt
Flask==1.1.2
Keras==2.1.1
tensorflow==1.14.0
matplotlib==3.1.3
numpy==1.19.5
scikit-image==0.14.2
h5py==2.10.0
scipy==1.4.1
imgaug==0.4.0
ipython==7.15.0
Pillow==8.4.0
eventlet==0.30.2
gunicorn==20.1.0
Flask-Testing==0.8.1
